+++ CbgTaxi24.API/Application/Queries/Dtos/RiderDtos.cs
@@ -8,16 +8,16 @@ namespace CbgTaxi24.API.Application.Queries.Dtos
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string OtherNames { get; set; }
-        public decimal Latitude { get; set; }
-        public decimal Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
         public string LocationName { get; set; }
     }
 
 
     public class LocationDto
     {
-        public decimal Latitude { get; set; }
-        public decimal Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
         //public string Region { get; set; }
         public string Name { get; set; }
     }
