+++ CbgTaxi24.API/Application/Services/RiderService.cs
@@ -1,4 +1,5 @@
 ﻿using CbgTaxi24.API.Application.Queries.Dtos;
+using CbgTaxi24.API.Application.Requests;
 using CbgTaxi24.API.Data;
 using CbgTaxi24.API.Infrastructure.Exceptions;
 using Microsoft.EntityFrameworkCore;
@@ -23,8 +24,8 @@ namespace CbgTaxi24.API.Application.Services
                                              RiderId = r.RiderId,
                                              FirstName = r.FirstName,
                                              LastName = r.LastName,
-                                             Latitude = r.Location.Latitude,
-                                             Longitude = r.Location.Longitude,
+                                             Latitude = (double)r.Location.Latitude,
+                                             Longitude = (double)r.Location.Longitude,
                                              LocationName = r.Location.Name
                                          })
                                          .ToListAsync();
@@ -48,8 +49,8 @@ namespace CbgTaxi24.API.Application.Services
                                         RiderId = r.RiderId,
                                         FirstName = r.FirstName,
                                         LastName = r.LastName,
-                                        Latitude = r.Location.Latitude,
-                                        Longitude = r.Location.Longitude,
+                                        Latitude = (double)r.Location.Latitude,
+                                        Longitude = (double)r.Location.Longitude,
                                         LocationName = r.Location.Name
                                     })
                                     .FirstOrDefaultAsync() ?? throw new PlatformException("not found");
